Renamed test-http-agent-maxsockets-regress-4050 to test-http-agent-maxsockets-respected
